backend/requirements.txt
Flask>=2.3.2
flask-cors>=4.0.0
requests>=2.31.0
transformers>=4.21.0
torch>=1.12.0
scikit-learn>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1